features: сделаны блоками все части фильтров, сделан второй тип фильтра админа
