+++ src/features/tasks-filter/ui/categories-block/index.tsx
@@ -5,17 +5,17 @@ import { FilterItemsIds } from "../consts";
 import styles from "../styles.module.css";
 
 interface Props {
-  filter: string[];
+  selectedCategories: string[];
   onChange: (name: string, value: string[]) => void;
 }
 
-export const CategoriesBlock = ({ filter, onChange }: Props) => {
+export const CategoriesBlock = ({ selectedCategories, onChange }: Props) => {
   const handleCheckboxChange = ({ target }: ChangeEvent<HTMLInputElement>) => {
     let newValue;
     if (target.checked) {
-      newValue = [...filter, target.id];
+      newValue = [...selectedCategories, target.id];
     } else {
-      newValue = filter.filter((item) => item !== target.id);
+      newValue = selectedCategories.filter((item) => item !== target.id);
     }
     onChange("categories", newValue);
   };
@@ -36,13 +36,13 @@ export const CategoriesBlock = ({ filter, onChange }: Props) => {
         <div className={styles.row}>
           <Checkbox
             label="Категория 1"
-            checked={filter.includes(FilterItemsIds.CATEGORY_1)}
+            checked={selectedCategories.includes(FilterItemsIds.CATEGORY_1)}
             id={FilterItemsIds.CATEGORY_1}
             onChange={handleCheckboxChange}
           />
           <Checkbox
             label="Категория 2"
-            checked={filter.includes(FilterItemsIds.CATEGORY_2)}
+            checked={selectedCategories.includes(FilterItemsIds.CATEGORY_2)}
             id={FilterItemsIds.CATEGORY_2}
             onChange={handleCheckboxChange}
           />
@@ -50,13 +50,13 @@ export const CategoriesBlock = ({ filter, onChange }: Props) => {
         <div className={styles.row}>
           <Checkbox
             label="Категория 3"
-            checked={filter.includes(FilterItemsIds.CATEGORY_3)}
+            checked={selectedCategories.includes(FilterItemsIds.CATEGORY_3)}
             id={FilterItemsIds.CATEGORY_3}
             onChange={handleCheckboxChange}
           />
           <Checkbox
             label="Категория 4"
-            checked={filter.includes(FilterItemsIds.CATEGORY_4)}
+            checked={selectedCategories.includes(FilterItemsIds.CATEGORY_4)}
             id={FilterItemsIds.CATEGORY_4}
             onChange={handleCheckboxChange}
           />
@@ -64,13 +64,13 @@ export const CategoriesBlock = ({ filter, onChange }: Props) => {
         <div className={styles.row}>
           <Checkbox
             label="Категория 5"
-            checked={filter.includes(FilterItemsIds.CATEGORY_5)}
+            checked={selectedCategories.includes(FilterItemsIds.CATEGORY_5)}
             id={FilterItemsIds.CATEGORY_5}
             onChange={handleCheckboxChange}
           />
           <Checkbox
             label="Категория 6"
-            checked={filter.includes(FilterItemsIds.CATEGORY_6)}
+            checked={selectedCategories.includes(FilterItemsIds.CATEGORY_6)}
             id={FilterItemsIds.CATEGORY_6}
             onChange={handleCheckboxChange}
           />
