+++ src/features/tasks-filter/ui/volunteer-filter/activeMode/index.tsx
@@ -11,6 +11,6 @@ interface Props {
 
 export const ActiveMode = ({ filter, onChange, modeOfProfile }: Props) => (<>
     <SortByBlock filter={filter.sortBy} onChange={onChange} userRole="volunteer"/>
-    <CategoriesBlock filter={filter.categories} onChange={onChange}/>
+    <CategoriesBlock selectedCategories={filter.categories} onChange={onChange}/>
     <RadiusBlock filter={filter.searchRadius} onChange={onChange} modeOfProfile={modeOfProfile} />
   </>)
