+++ src/features/tasks-filter/ui/volunteer-filter/index.tsx
@@ -14,8 +14,7 @@ export const VolunteerFilter = ({ filter, onChange }: Props) => {
 
   // определяем текущий вариант профиля волонтёра по пути в URL
   const pathnameAsArray = useLocation().pathname.slice(1).split('/');
-  const indexOfVolunteerInArray = pathnameAsArray.indexOf('volunteer');
-  const currentModeOfProfile = pathnameAsArray[indexOfVolunteerInArray + 1];
+  const currentModeOfProfile = pathnameAsArray[pathnameAsArray.indexOf('volunteer') + 1];
 
   let currentModeOfFilter: JSX.Element;
   switch (currentModeOfProfile) {
